@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { OnBoardComputerContComponent } from "./on-board-computer-cont.component";
 import { OddMissionService } from "../../../services/odd-mission/odd-mission.service";
-import { HttpClientModule } from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import { OnBoardComputerPresComponent } from "./pres/on-board-computer-pres.component";
 
 describe("OnBoardComputerContComponent", () => {
@@ -10,12 +10,11 @@ describe("OnBoardComputerContComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [HttpClientModule],
       declarations: [
         OnBoardComputerContComponent,
         OnBoardComputerPresComponent
       ],
-      providers: [OddMissionService, OddMissionService]
+      providers: [OddMissionService, OddMissionService, provideHttpClient(withInterceptorsFromDi())]
     }).compileComponents();
 
     fixture = TestBed.createComponent(OnBoardComputerContComponent);
