@@ -1,14 +1,14 @@
 import { TestBed } from "@angular/core/testing";
 import { OddMissionService } from "./odd-mission.service";
-import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideHttpClientTesting } from "@angular/common/http/testing";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe("OddMissionService", () => {
   let service: OddMissionService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-      providers: [OddMissionService]
+      providers: [OddMissionService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
     });
     service = TestBed.inject(OddMissionService);
   });
